Main ewm service

+++ stat-service/stat-server/src/main/java/ru/practicum/ewm/mapper/ViewStatMapper.java
@@ -16,7 +16,7 @@ public class ViewStatMapper implements RowMapper<ViewStats> {
         return ViewStats.builder()
                 .app(rs.getString("app"))
                 .uri(rs.getString("uri"))
-                .hits(rs.getInt("hits"))
+                .hits(rs.getLong("hits"))
                 .build();
     }
 }
